Create Second Commit
With this commit, my focus was on making the multiplication and division segments multi entry. This proved challenging, as it was required to remember the previous loop response and carry on in such a way. Once I thought through the logic of how a calculator would interpret this, I managed to get it down.