Create Third Commit
With this commit, I wanted to give the user more options when it came to percentages, as I feel limiting the process to just finding the percentage of a number doesn't provide as much choice as I would like to exhibit. As my whole focus for this project is to develop my skills with Python, I wanted to challenge myself to add this extra option under a predefined choice.
Again, I found it useful to call upon algebraic methods to clearly lay out how an individual would process such a question. I have found that using Mathematical functions has greatly helped me with this project in general, and I am excited to see where else I am able to apply this knowledge.
I found it quite difficult to word the inputs, and therefore added a brief template using the print() function in order to enable the user to read the process clearly, and understand how to input their request.